backend/requirements.txt
python-dotenv
flask
flask_limiter
flask_cors
gunicorn
